Fixed datatype

+++ analysis/RDF/RDF.py
@@ -62,7 +62,7 @@ def rdf(x, t, box, pairs, nbins, pbc=True):
   natoms = np.shape(x)[0]
   npairs = len(pairs)
   ncol = npairs + 1
-  pair_ar = np.zeros(2*npairs)
+  pair_ar = np.zeros(2*npairs, dtype=np.int32)
   i = 0
   for p in pairs:
     pair_ar[i] = p[0][0]
